Cast result to bigint (PostgreSQL 12)

+++ FUNCTIONS/round_to_sig_figs.sql
@@ -9,5 +9,5 @@ CREATE OR REPLACE FUNCTION timeit.round_to_sig_figs(bigint, integer)
 RETURNS bigint
 LANGUAGE sql
 AS $$
-SELECT round($1, $2 - 1 - floor(coalesce(log(nullif(abs($1),0)),0))::int);
+SELECT round($1, $2 - 1 - floor(coalesce(log(nullif(abs($1),0)),0))::int)::bigint;
 $$;
